client protocol init
