+++ src/py/flwr/client/app_test.py
@@ -17,7 +17,6 @@
 
 from typing import Dict, Tuple
 
-from flwr.client import ClientLike, to_client
 from flwr.common import (
     Config,
     EvaluateIns,
@@ -83,26 +82,19 @@ class NeedsWrappingClient(NumPyClient):
 
 def test_to_client_with_client() -> None:
     """Test to_client."""
-    # Prepare
-    client_like: ClientLike = PlainClient()
-
-    # Execute
-    actual = to_client(client_like=client_like)
+    client = PlainClient().to_client()
 
     # Assert
-    assert isinstance(actual, Client)
+    assert isinstance(client, Client)
 
 
 def test_to_client_with_numpyclient() -> None:
     """Test fit_clients."""
-    # Prepare
-    client_like: ClientLike = NeedsWrappingClient()
-
     # Execute
-    actual = to_client(client_like=client_like)
+    client = NeedsWrappingClient().to_client()
 
     # Assert
-    assert isinstance(actual, Client)
+    assert isinstance(client, Client)
 
 
 def test_start_client_transport_invalid() -> None:
