+++ src/py/flwr/client/client_protocol.py
@@ -1,4 +1,4 @@
-# Copyright 2020 Adap GmbH. All Rights Reserved.
+# Copyright 2023 Flower Labs GmbH. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,23 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Flower client."""
+"""Client protocols."""
+
+from typing import Protocol
+
+from flwr.client import Client
 
 
-from .app import start_client as start_client
-from .app import start_numpy_client as start_numpy_client
-from .client import Client as Client
-from .numpy_client import NumPyClient as NumPyClient
-from .run import run_client as run_client
-from .typing import ClientFn as ClientFn
-from .typing import ClientLike as ClientLike
+class ClientProtocol(Protocol):
+    """Define a Protocol for to associate client functionality."""
 
-__all__ = [
-    "Client",
-    "ClientFn",
-    "ClientLike",
-    "NumPyClient",
-    "run_client",
-    "start_client",
-    "start_numpy_client",
-]
+    def to_client(self) -> Client:
+        """Convert any client type to Client."""
