+++ src/py/flwr/simulation/ray_transport/ray_client_proxy.py
@@ -22,7 +22,7 @@ from typing import Dict, Optional, cast
 import ray
 
 from flwr import common
-from flwr.client import Client, ClientFn, ClientLike, to_client
+from flwr.client import Client, ClientFn
 from flwr.client.client import (
     maybe_call_evaluate,
     maybe_call_fit,
@@ -275,5 +275,4 @@ def launch_and_evaluate(
 
 def _create_client(client_fn: ClientFn, cid: str) -> Client:
     """Create a client instance."""
-    client_like: ClientLike = client_fn(cid)
-    return to_client(client_like=client_like)
+    return client_fn(cid).to_client()
